% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_Z.R
\name{bootstrap_Z}
\alias{bootstrap_Z}
\title{title.}
\usage{
bootstrap_Z(df, value, est, by = NULL, Z_est = "Z_est",
  Z_value = "Z_value")
}
\arguments{
\item{df}{A dataframe}

\item{value}{The (quoted) name of the column in df with bootstrapped estimates}

\item{est}{The (quoted) name of the column in df with observed estimates, for which Z-scores should be calculated}

\item{by}{An option vector of (quoted) column names to group df by. Means and standard deviations for the Z scores will be calculated within these groups. If by=NULL (default), then no grouping is used.}

\item{Z_est}{The (quoted) name of the column to create for the Z scores from est}

\item{Z_value}{The (quoted) name of the column to create for the Z scores from value}
}
\value{
A dataframe with columns for each grouping variable in by (if used), est, value, mean and sd from value, and Z scores for the est and value columns.
}
\description{
Calculates the Z-scores for observed estimates based on the mean and SD of a bootstrapped null distribution.
Requires a dataframe in long format, with bootstrapped estimates in value and observed estimates 
(for which Z-scores should be calculated) in est.
}
\examples{

 
}
\author{
Rose Hartman
}

