% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_word}
\alias{convert_to_word}
\title{Convert to word}
\usage{
convert_to_word(x, cap = FALSE)
}
\arguments{
\item{x}{An integer 0-9}

\item{cap}{Should the first letter be capitalized? Default is FALSE.}
}
\value{
Returns the word form of x (e.g. 0 returns "zero").
}
\description{
Converts single digit numbers (0-9) into words, for use in manuscripts
}

