% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_context_vs_nontext.R
\name{plot_context_vs_nontext}
\alias{plot_context_vs_nontext}
\title{Creates a boxplot showing bootstrapped null distirbutions with context estimates as colored dots.}
\usage{
plot_context_vs_nontext(results, outcome, xlabs = FALSE, Z.score = FALSE,
  methods_wrap = FALSE, save.to, ...)
}
\arguments{
\item{argument_name}{description of the argument}

\item{argument_name}{description of the argument}

\item{argument_name}{description of the argument}

\item{argument_name}{description of the argument}

\item{argument_name}{description of the argument}
}
\value{
None.
}
\description{
Longer description and notes.
}
\examples{
plot_context_vs_nontext(context_all_freq, nontext_results_freq, global_results_freq, outcome="precision", xlabs=TRUE, save.to="plots/context_vs_nontext", consider.freq=TRUE)

}
\author{
Rose Hartman
}

