% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_for_cm.R
\name{print_for_cm}
\alias{print_for_cm}
\title{Prepare transcript files for use by computational models.}
\usage{
print_for_cm(df, nontext, dir = getwd(), save.to = "computational_models",
  r = 0, quiet = TRUE)
}
\description{
For use with dpseg-1.2.1, corpus should be represented in the following way:
one utterance per line
a space bewteen each word
words represented with dictionary pronunciations with one character per phoneme
}
\details{
For use with py-cfg via colingFinal, corpus should be represented in the following way:
one utterance per line
spaces between each phoneme
tabs between each word
words represented with dictionary pronunciations using a limited set of possible characters (just alpha, I think). A single phoneme may be represented with more than one character.
}

