% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_for_cm.R
\name{print_for_coling}
\alias{print_for_coling}
\title{Print for coling computational model}
\usage{
print_for_coling(df, nontext, dir = getwd(),
  save.to = "computational_models", r = 0, quiet = TRUE)
}
\description{
Prepares (saves to text files) phonetic approximations of transcript utterances for use by coling computational models.
}

