% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_convergence.R
\name{resampling_convergence}
\alias{resampling_convergence}
\title{Save plots showing boostrapped trajectories for resampled estimates}
\usage{
resampling_convergence(samples, n = 100, save.to = NULL, measure = NULL,
  note = NULL, plot.mean = TRUE, plot.sig.cutoffs = TRUE, tol = 1)
}
\arguments{
\item{samples}{The vector of samples from the resampling procedure}

\item{n}{The number of lines (bootstrapped trajectories) to plot}

\item{save.to}{directory for saving plots. If NULL, no plots will be saved.}

\item{measure}{a character string describing the estimates. This will be included in the file names (if saving plots) and as the label on the y-axis of the plots.}

\item{note}{a character string providing any additional information that should be saved in the file name}

\item{plot.mean}{produce a plot showing the bootstrapped trajectories for the mean?}

\item{plot.sig.cutoffs}{produce a plot showing the bootstrapped trajectories for the 97.5th and 2.5th percentiles (the cutoffs for significance at alpha=.05 for two-tailed tests)?}

\item{tol}{the half width of the desired stability corridor. To not plot the stability corridor, set tol=0.}
}
\description{
Saves plots that show n bootstrapped trajectories for estimates from a resampling procedure. 
Includes horizontal reference lines for the observed estimate (mean of all resampled estimates) and a corridor of stability +-tol from that estimate.
}

