% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{within_k}
\alias{within_k}
\title{Tests for lag (or lead) within k spaces away}
\usage{
within_k(x, k, type)
}
\arguments{
\item{x}{a vector where entries are all 1's or 0's (or NA's)}

\item{k}{the distance over which to calculate lag or lead}

\item{type}{should be "lag" or "lead", which call \code{\link[dplyr]{lag}} or
\code{\link[dplyr]{lead}} respectively, or "either", which calls both.}
}
\value{
A vector of the same length as x, indicating whether there is a 1 in 
  x anywhere up to k entries before (for type = "lag"), after (for type = 
  "lead") or in either direction (for type ="either"). Will not return NAs.
  Note that if x contains numbers other than 0 and 1, it will just treat 1 as
  a success and all other values (including NA) as failure.
}
\description{
Tests for lag (or lead) within k spaces away
}
\examples{
# quarters in which presidental approval dropped below 50\% 
x <- as.numeric(presidents < 50) 
within_k(x, k=4, "lag") # quarters that follow quarters with approval below 50\% by 1 to 4 quarters 
within_k(x, k=4, "either") # quarters that fall within 4 quarters from one with an approval below 50\%

}

